requirements.txt
# Processing HTTP requests
requests

# Read YAML file
ruamel.yaml

# Data Processing
pandas

# FastAPI requirements
fastapi
pydantic
uvicorn
